+++ plugins/pug/src/services/documentHighlight.ts
