+++ plugins/pug/src/services/documentLinks.ts
