+++ plugins/pug/src/services/documentSymbol.ts
