+++ plugins/pug/src/services/foldingRanges.ts
