+++ plugins/pug/src/services/hover.ts
