+++ plugins/pug/src/services/scanner.ts
