--- pug-language-tools/pug-language-service/tsconfig.build.json
@@ -1,26 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"noEmit": false,
-		"outDir": "out",
-		"rootDir": "src",
-	},
-	"include": [
-		"src"
-	],
-	"exclude": [
-		"node_modules",
-		".vscode-test",
-	],
-	"references": [
-		{
-			"path": "../../packages/shared/tsconfig.build.json"
-		},
-		{
-			"path": "../../packages/source-map/tsconfig.build.json"
-		},
-		{
-			"path": "../../packages/language-service/tsconfig.build.json"
-		},
-	],
-}
